Move xen storage tests to own file
This allows to run just the general storage tests without all the slow xen
storage based tests.
