+++ tests/__init__.py
@@ -584,6 +584,7 @@ def load_tests(loader, tests, pattern):
             'qubes.tests.backupcompatibility',
             'qubes.tests.regressions',
             'qubes.tests.storage',
+            'qubes.tests.storage_xen',
             ):
         tests.addTests(loader.loadTestsFromName(modname))
